+++ SmartApp.AV/UI/src/ui/colorful/widget/svgface/IAnimated.ts
@@ -1,0 +1,3 @@
+export interface IAnimated {
+    update(updates: object): void;
+}
